print data table and show in UI to confirm state of bill
Tab rows are only working with one row.
It overwrites the previous row.
So tabs array will always be length 1 with the most recently edited tab
in the array
The bill will not always output correct state (i.e. one keystroke behind)
This may be because of setShowBill state though